Extract MaterialFullScreenDialogFragment
