+++ collect_app/src/main/res/layout/identify_user_dialog.xml
@@ -7,26 +7,7 @@
     android:background="?attr/colorSurface"
     android:orientation="vertical">
 
-    <com.google.android.material.appbar.AppBarLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <com.google.android.material.appbar.MaterialToolbar
-            android:id="@+id/toolbar"
-            android:layout_width="match_parent"
-            android:layout_height="?attr/actionBarSize"
-            app:navigationContentDescription="@string/close"
-            app:navigationIcon="@drawable/ic_close_black_24dp"
-            tools:title="Form Name" />
-
-        <include
-            android:id="@+id/action_bar_shadow"
-            layout="@layout/toolbar_action_bar_shadow"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:visibility="gone" />
-
-    </com.google.android.material.appbar.AppBarLayout>
+    <include layout="@layout/material_full_screen_dialog_toolbar" />
 
     <ScrollView
         android:layout_width="match_parent"
